C++ Implementation of Wiimote Controller Node
Rewrite of the Wiimote node in C++ using the existing Python node
as the requirements. The C++ implementation was design with focus
on reduced resource consumption.

Using C++ cut the memory consumption by 50-60%, and the reduced the
number of cpu-clock cycles used 20x-30x depending upon the use case.

Differences from Python implementation:
- Both "/wiimote/nunchuk" and "/wiimote/classic" topics are only
  published if the Nunchuk or Classic Controller are connected to
  the wiimote respectively.
- Wiimote data is only polled from the controller if the data is
  required to publish for a topic which has at least one subscriber.